+++ src/main/kotlin/goodspace/bllsoneshot/task/dto/response/submit/ProofShotSubmitResponse.kt
@@ -1,4 +1,4 @@
-package goodspace.bllsoneshot.task.dto.response
+package goodspace.bllsoneshot.task.dto.response.submit
 
 data class ProofShotSubmitResponse(
     val imageFileId: Long,
