+++ src/main/kotlin/goodspace/bllsoneshot/task/dto/response/submit/QuestionSubmitResponse.kt
@@ -1,4 +1,4 @@
-package goodspace.bllsoneshot.task.dto.response
+package goodspace.bllsoneshot.task.dto.response.submit
 
 data class QuestionSubmitResponse(
     val number: Int,
