+++ src/main/kotlin/goodspace/bllsoneshot/task/dto/response/submit/TaskSubmitResponse.kt
@@ -1,4 +1,4 @@
-package goodspace.bllsoneshot.task.dto.response
+package goodspace.bllsoneshot.task.dto.response.submit
 
 import goodspace.bllsoneshot.entity.assignment.Subject
